EEBus: keep optional capabilities in base implementations
